comments and duplicate code
reduced comment size and deleted a lot of them and I believe i removed the correct section of duplicate code - there was an outside repository under shared called "game" that was not under chess, if it tells me the files are missing I will reverse the change to no longer have game under chess. here goes.