ok SO
The way you have to put the WS BEFORE the initialization is bogglign to the brain and my boy chat helped me figure out the format which was super frustrating, BUT! we now know where to put him. we can pass things back and forth. time to figure out how to pass messages back and forth.